requirements-test.txt
bcrypt==3.1.5
bleach==3.0.2
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
docutils==0.14
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
flask-session-plus==0.0.2
Flask-WTF==0.14.2
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
googleapis-common-protos==1.5.5
grpcio==1.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pkginfo==1.4.2
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
Pygments==2.3.1
pytz==2018.7
readme-renderer==24.0
requests==2.21.0
requests-toolbelt==0.8.0
rsa==4.0
six==1.12.0
tqdm==4.28.1
twine==1.12.1
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.2.1
